Rename map layout into suivi-pcrs-map-layout
